Added Category model and controller.

+++ src/main/java/com/dawidp/warehousemanagementsystem/controller/CustomerController.java
@@ -24,7 +24,7 @@ public class CustomerController {
 	}
 
 	@GetMapping("/getCustomer/{customerId}")
-	public Customer getCustomer(@PathVariable int customerId) {
+	public Customer getCustomer(@PathVariable Long customerId) {
 		return service.getCustomer(customerId);
 	}
