+++ src/main/java/com/dawidp/warehousemanagementsystem/dao/CustomerRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.dawidp.warehousemanagementsystem.model.Customer;
 
 @Repository
-public interface CustomerRepository extends JpaRepository<Customer, Integer> {
+public interface CustomerRepository extends JpaRepository<Customer, Long> {
 
-	Customer findCustomerByCustomerId(int customerId);
+	Customer findCustomerByCustomerId(Long customerId);
 }
